Bishop moves
In chess, the bishop moves diagonally, any number of squares. Given two different squares of the chessboard, determine whether a bishop can go from the first to the second in one move. The program should output YES if a Bishop can go from the first square to the second in one move, or NO otherwise.